requirements.txt
requests
python-telegram-bot
# pip install --upgrade apscheduler pytz tzlocal

google-cloud-storage
google-cloud-firestore
google-cloud-pubsub

flask
paho-mqtt
flask-mqtt

numpy
pandas
matplotlib
scikit-learn
plotly

google-cloud-aiplatform
google-cloud-texttospeech
google-genai

